Adding VRC7 support, including emu2413
One of the reasons VRC7 (and S5B) wasn't included earlier was because of potential
copyright/licensing issues. emu2413.c doesn't explicitly include a copyright license
(be it LGPL, MIT, BSD, public domain, etc.).
The source of the file originates from the MAME project, in which developers may individually
license their contributions.
The license (zlib) can be found: http://mednafen.sourceforge.net/documentation/09x/mednafen.html
I believe Mednafen to be a good authority on the matter, and Google seems to agree with the consensus
regarding the license for emu2413.c

+++ src/famitracker-core/APU/VRC7.h
@@ -31,7 +31,7 @@ public:
 
 	void Reset();
 	void SetSampleSpeed(uint32 SampleRate, double ClockRate, uint32 FrameRate);
-	void SetVolume(int Volume);
+	void SetVolume(float Volume);
 	void Write(uint16 Address, uint8 Value);
 	uint8 Read(uint16 Address, bool &Mapped);
 	void EndFrame();
